Mirror the EVM util layout: split svm into svm/svm_tokens/svm_transaction
Match core/utils/{web3,tokens,transaction}.py:
- svm.py            -> RPC client resolution (mirrors web3.py)
- svm_tokens.py     -> native/SPL balances + mint decimals (mirrors tokens.py)
- svm_transaction.py-> broadcast + confirm (mirrors transaction.py)

Repoint token_resolver + tests; all 36 Solana tests pass.
